Consolidate to one single pass
