+++ examples/real_dims2.py
@@ -2,7 +2,7 @@ import torch
 import torch_spyre._inductor.passes as passes
 
 # import torch_spyre._inductor.propagate_real_dims as prd
-import torch_spyre._inductor.propagate_real_dims2 as prd
+import torch_spyre._inductor.propagate_real_dims as prd
 
 passes.propagate_real_dims = prd.propagate_real_dims
 declare_real_dim = prd.declare_real_dim
