+++ torch_spyre/_inductor/propagate_real_dims.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 
+import sympy
 import torch
 from .logging_utils import get_inductor_logger
 from torch._inductor.ir import (
@@ -30,7 +31,9 @@ from torch._inductor.dependencies import MemoryDep
 from torch._inductor.virtualized import V
 from .errors import Unsupported
 from .constants import BATCH_MATMUL_OP
-from .views import compute_coordinates
+from .pass_utils import host_coordinates, device_coordinates
+from .views import matching_dim, compute_coordinates
+from torch_spyre._C import SpyreTensorLayout
 from torch.utils.weak import WeakTensorKeyDictionary
 
 logger = get_inductor_logger("propagate_real_dims")
@@ -60,10 +63,12 @@ def _get_buffer(dep):
     return V.graph.get_buffer(dep.name)
 
 
+def _lone_sym(coord: sympy.Expr) -> sympy.Symbol:
+    return next(iter(coord.free_symbols))
+
+
 def _compute_real_layout(real_dims):
-    """
-    Compute real size and stride based on real dims
-    """
+    """Compute real size and stride from declared real dim sizes."""
     size = []
     stride = [1]
     for s in reversed(real_dims):
@@ -72,58 +77,96 @@ def _compute_real_layout(real_dims):
     return list(reversed(size)), list(reversed(stride[:-1]))
 
 
-def _matmul_real_dims(op, inputs):
+def compute_input_real_dims(dep: MemoryDep) -> dict:
+    """Map loop vars to real dim names for a single input dep, using real-space coords."""
+    buf = _get_buffer(dep)
+    if not hasattr(buf, 'real_dims') or buf.real_dims is None:
+        # Scalar broadcast: constant index, contributes nothing to rdims
+        if not dep.index.free_symbols:
+            return {}
+        return None
+    real_size, real_stride = _compute_real_layout(buf.real_dims)
+    coords = compute_coordinates(real_size, real_stride, dep.ranges, dep.index)
+    result = {}
+    for i, coord in enumerate(coords):
+        if coord.free_symbols:
+            sym = _lone_sym(coord)
+            name = buf.real_dims[i]
+            actual_range = int(dep.ranges[sym])
+            declared_size = _real_dims.get(name)
+            if declared_size is not None and actual_range != declared_size:
+                logger.warning(
+                    f"{dep.name}: loop var {sym} has range {actual_range} "
+                    f"but maps to '{name}' declared as {declared_size} -- partial/sliced dim"
+                )
+            result.setdefault(sym, []).append(name)
+    return result
+
+
+def op_out_coords(op: ComputedBuffer) -> list:
+    output_dep = next(iter(op.get_read_writes().writes))
+    return host_coordinates(op.get_layout(), output_dep)
+
+
+def coords_to_real_dims(coords: list, rdims: dict) -> list:
+    """Map coordinate expressions to real dim names via their loop variable."""
+    result = []
+    for c in coords:
+        if c.free_symbols:
+            sym = _lone_sym(c)
+            if sym not in rdims:
+                logger.warning(f"coords_to_real_dims: no mapping for {sym} -- returning None")
+                return None
+            result.extend(rdims[sym])
+    return result
+
+
+def get_input_real_dims(inputs: list) -> dict:
     """
-    Augment matmul with real ranges and output real dims
+    Manage real_dims for all inputs, mapping dims from upstream node to this node.
+    Returns None if any input has no real_dims.
     """
-    ranges0 = inputs[0].ranges
-    index0 = inputs[0].index
-    real_dims0 = _get_buffer(inputs[0]).real_dims
-    real_size0, real_stride0 = _compute_real_layout(real_dims0)
- 
-    ranges1 = inputs[1].ranges
-    index1 = inputs[1].index
-    real_dims1 = _get_buffer(inputs[1]).real_dims
-    real_size1, real_stride1 = _compute_real_layout(real_dims1)
+    rdims = {}
+    for inp in inputs:
+        new = compute_input_real_dims(inp)
+        if new is None:
+            return None
+        rdims.update(new)
+    return rdims
 
-    # compute coordinates based on real dims
-    real_coords0 = compute_coordinates(real_size0, real_stride0, ranges0, index0)
-    real_coords1 = compute_coordinates(real_size1, real_stride1, ranges1, index1)
 
-    # identify order of op dims
-    keys = ranges0.keys()
-    vars0 = index0.free_symbols
-    vars1 = index1.free_symbols
-    dim0 = next(iter(keys - vars1))
-    dim1 = next(iter(vars0 & vars1))
-    dim2 = next(iter(keys - vars0))
+def get_reduction_dim(dep: MemoryDep, out_coords: list) -> sympy.Symbol:
+    """Return the reduction loop variable: the input coord absent from the output."""
+    in_coords = host_coordinates(_get_buffer(dep).get_layout(), dep)
+    reduction_coord = next(
+        c for c in in_coords
+        if c.free_symbols and matching_dim(out_coords, c) is None
+    )
+    return _lone_sym(reduction_coord)
 
-    # match iteration variables to real dims
-    matches = {dim0: None, dim1: None, dim2: None}
-    for i, coord in enumerate(real_coords0):
-        if coord.is_symbol:
-            matches[next(iter(coord.free_symbols))] = real_dims0[i]
-    for i, coord in enumerate(real_coords1):
-        if coord.is_symbol:
-            matches[next(iter(coord.free_symbols))] = real_dims1[i]
 
-    # add output real dims
-    op.real_ranges = [matches[dim0], matches[dim1], matches[dim2]]
-
-    # add op real ranges
-    op.real_dims = [matches[dim0], matches[dim2]]
+def _set_no_real_dims(op):
+    op.real_dims = None
+    op.reduction_dims = None
+    op.rdims = {}
 
 
 def _compute_real_dims(op, inputs):
-    """
-    Augment op with real ranges and output real dims
-    """
-    if isinstance(op.data, Reduction) and op.data.reduction_type == BATCH_MATMUL_OP:
-        return _matmul_real_dims(op, inputs)
-
-    # TODO handle other op types
-    op.real_ranges = _get_buffer(inputs[0]).real_dims
-    op.real_dims = _get_buffer(inputs[0]).real_dims
+    rdims = get_input_real_dims(inputs)
+    if rdims is None:
+        _set_no_real_dims(op)
+        return
+    out_coords = op_out_coords(op)
+    real_dims = coords_to_real_dims(out_coords, rdims)
+    if real_dims is None:
+        _set_no_real_dims(op)
+        return
+    op.real_dims = real_dims
+    op.rdims = rdims
+    if isinstance(op.data, Reduction):
+        op.reduction_dims = rdims[get_reduction_dim(inputs[0], out_coords)]
+    else:
+        op.reduction_dims = None
 
 
 def propagate_real_dims(
@@ -149,6 +192,20 @@ def propagate_real_dims(
                     raise Unsupported(f"graph input {name} does not have a FixedLayout")
                 tb.real_dims = _real_tensor_dims.get(real_input)
 
+    def _dump_dep(label, dep):
+        buf = V.graph.get_buffer(dep.name)
+        layout = buf.get_layout() if hasattr(buf, 'get_layout') else None
+        real_dims = getattr(buf, 'real_dims', '?')
+        logger.debug(f"  {label} {dep.name}: real_dims={real_dims}")
+        if layout is not None:
+            logger.debug(f"    host_size={list(layout.size)}  host_stride={list(layout.stride)}")
+            logger.debug(f"    host_coordinates={host_coordinates(layout, dep)}")
+        stl = getattr(buf, 'layout', None)
+        if isinstance(stl, SpyreTensorLayout):
+            logger.debug(f"    device_size={stl.device_size}  stride_map={stl.stride_map}")
+            logger.debug(f"    device_coordinates={device_coordinates(stl, dep)}")
+        logger.debug(f"    index={dep.index}  ranges={dict(dep.ranges)}")
+
     it = iter(operations)
     for op in it:
         if op.is_no_op():
@@ -157,33 +214,70 @@ def propagate_real_dims(
         elif isinstance(op, ComputedBuffer):
             if isinstance(op.layout, MutationLayoutSHOULDREMOVE):
                 continue
+            origins = getattr(op.data, 'origins', set())
+            aten_ops = [str(n.target) for n in origins if hasattr(n, 'target')]
+            reduction_type = getattr(op.data, 'reduction_type', None)
+            logger.debug(f"\n--- {op.get_operation_name()} ({type(op.data).__name__}) aten={aten_ops} reduction_type={reduction_type}")
             rw = op.get_read_writes()
             inputs = []
             for input in rw.reads:
                 if isinstance(input, MemoryDep):
                     inputs.append(input)
+                    _dump_dep("input", input)
+            for write in rw.writes:
+                if isinstance(write, MemoryDep):
+                    _dump_dep("output", write)
             if isinstance(op.data, (Pointwise, Reduction)):
-                if any(_get_buffer(inp).real_dims is None for inp in inputs):
-                    logger.warning(f"skipping {op.get_operation_name()}: input has no real_dims")
-                    op.real_dims = None
-                    op.real_ranges = None
-                else:
-                    _compute_real_dims(op, inputs)
+                _compute_real_dims(op, inputs)
             else:
                 logger.warning(f"Warning: unhandled node type {type(op.data)}")
-                op.real_dims = None
-                op.real_ranges = None
+                _set_no_real_dims(op)
         else:
             logger.warning(f"unhandled operation type {type(op)}")
-            op.real_dims = None
-            op.real_ranges = None
+            _set_no_real_dims(op)
 
     # debug
 
+    print("DECLARED DIMS")
+    for name, size in _real_dims.items():
+        print(f"  {name} = {size}")
+
+    print("INPUT TENSORS")
+    for name in V.graph.graph_input_names:
+        tb = V.graph.graph_inputs[name]
+        if isinstance(tb, TensorBox):
+            print(f"  {name}: real_dims={tb.real_dims}")
+
     print("OPS")
     for op in iter(operations):
-        print(op.get_operation_name(), op.real_ranges)
-
-    print("TENSORS")
-    for buf in V.graph.buffers:
-        print(buf.name, buf.real_dims)
+        if not hasattr(op, 'rdims') or op.real_dims is None:
+            origins = getattr(getattr(op, 'data', op), 'origins', set())
+            aten_ops = [str(n.target) for n in origins if hasattr(n, 'target')]
+            print(f"  {op.get_operation_name()}: skipped ({type(op).__name__} / {type(getattr(op, 'data', op)).__name__})  aten={aten_ops}")
+            continue
+        is_reduction = isinstance(op.data, Reduction)
+        origins = getattr(op.data, 'origins', set())
+        aten_ops = [str(n.target) for n in origins if hasattr(n, 'target')]
+        reduction_type = getattr(op.data, 'reduction_type', None)
+        print(f"  {op.get_operation_name()} ({'reduction' if is_reduction else 'pointwise'})  aten={aten_ops}  reduction_type={reduction_type}")
+        rw = op.get_read_writes()
+        all_deps = list(rw.reads) + list(rw.writes)
+        for dep in rw.reads:
+            if isinstance(dep, MemoryDep):
+                buf = _get_buffer(dep)
+                real_dims = getattr(buf, 'real_dims', '?')
+                host_size = list(buf.get_layout().size) if hasattr(buf, 'get_layout') else '?'
+                print(f"    input {dep.name}: real_dims={real_dims}  host_size={host_size}  index={dep.index}  ranges={dict(dep.ranges)}")
+        print(f"    loop vars:")
+        ranges = {}
+        for dep in all_deps:
+            if isinstance(dep, MemoryDep):
+                ranges.update({str(s): int(v) for s, v in dep.ranges.items()})
+        for sym, names in op.rdims.items():
+            size = ranges.get(str(sym), "?")
+            declared = [f"{n}={_real_dims.get(n,'?')}" for n in names]
+            print(f"      {sym}: range={size}  real_dim(s)={names}  declared={declared}")
+        if is_reduction:
+            print(f"    reduction over: {op.reduction_dims}")
+        print(f"    output: ({op.get_name()}) real_dims={op.real_dims}")
+        print()
